Rename generate_numeric_debug_handle to numeric_debugger
Summary:
att

Test Plan:
CI

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
