+++ docs/source/quantization-support.rst
@@ -635,4 +635,4 @@ the `custom operator mechanism <https://pytorch.org/tutorials/advanced/torch_scr
 .. automodule:: torch.nn.quantized.dynamic.modules
 .. automodule:: torch.quantization
 .. automodule:: torch.nn.intrinsic.modules
-.. automodule:: torch.ao.quantization.pt2e.generate_numeric_debug_handle
+.. automodule:: torch.ao.quantization.pt2e.numeric_debugger
