+++ torch/ao/quantization/fx/convert.py
@@ -75,7 +75,7 @@ from .custom_config import (
 from .lower_to_fbgemm import lower_to_fbgemm
 # importing the lib so that the quantized_decomposed ops are registered
 from ._decomposed import quantized_decomposed_lib  # noqa: F401
-from torch.ao.quantization.pt2e.generate_numeric_debug_handle import NUMERIC_DEBUG_HANDLE_KEY
+from torch.ao.quantization.pt2e.numeric_debugger import NUMERIC_DEBUG_HANDLE_KEY
 import operator
 
 __all__ = [
